requirements.txt
# Core
google-genai>=1.47.0

# Web scraping
crawl4ai>=0.3.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Environment
python-dotenv>=1.0.0
